Add EndPoints Saque, Deposito, Ver Saldo

+++ bank/src/main/java/com/api/bank/BankApiLayer/Repository/TransactionHistoryDataRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionHistoryDataRepository extends JpaRepository<TransactionHistoryData, Integer> {
+public interface TransactionHistoryDataRepository extends JpaRepository<TransactionHistoryData, Long> {
     TransactionHistoryData findByCpf(String cpf); // Método para buscar por CPF
 
     void deleteByCpf(String cpf); // Método para excluir por CPF
